@@ -1,51 +1,44 @@
 {
-  "name": "miia-mini",
-  "description": "A PWA client for Hydrus",
-  "version": "0.1.0",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dpongimo/miia-web.git"
-  },
-  "scripts": {
-    "dev": "sapper dev",
-    "build": "sapper build --legacy",
-    "export": "sapper export --legacy --entry \"/ /tags\"",
-    "start": "node __sapper__/build"
-  },
-  "dependencies": {
-    "bootstrap": "^5.0.0-alpha1",
-    "compression": "^1.7.4",
-    "dexie": "^3.0.1",
-    "lightgallery.js": "^1.2.0",
-    "polka": "next",
-    "sirv": "^1.0.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.10.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/runtime": "^7.10.4",
-    "@rollup/plugin-babel": "^5.0.4",
-    "@rollup/plugin-commonjs": "^13.0.0",
-    "@rollup/plugin-node-resolve": "^8.1.0",
-    "@rollup/plugin-replace": "^2.3.3",
-    "@sveltejs/svelte-virtual-list": "^3.0.1",
-    "async-hydrus.js": "dpongimo/async-hydrus.js",
-    "autoprefixer": "^9.8.4",
-    "node-sass": "^4.14.1",
-    "npm-run-all": "^4.1.5",
-    "rollup": "^2.19.0",
-    "rollup-plugin-svelte": "^5.2.3",
-    "rollup-plugin-terser": "^6.1.0",
-    "sapper": "^0.27.16",
-    "svelte": "^3.23.2",
-    "svelte-lazy": "^0.1.13",
-    "svelte-preprocess": "^3.9.12",
-    "typescript": "^4.0.2"
-  },
-  "_removedDeps": {
-    "async-hydrus.js": "dpongimo/async-hydrus.js"
-  }
+	"name": "miia-mini",
+	"description": "A PWA client for Hydrus",
+	"version": "0.2.0",
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/dpongimo/miia-web.git"
+	},
+	"private": true,
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --plugin-search-dir . --check .",
+		"format": "prettier --plugin-search-dir . --write ."
+	},
+	"dependencies": {
+		"@dpongimo/async-hydrus.js": "^2.0.0-alpha5",
+		"bootstrap": "^5.2.2",
+		"dexie": "^3.2.2",
+		"lightgallery.js": "^1.4.0"
+	},
+	"devDependencies": {
+		"@sveltejs/adapter-static": "1.0.0-next.46",
+		"@sveltejs/kit": "1.0.0-next.522",
+		"@sveltejs/svelte-virtual-list": "^3.0.1",
+		"prettier": "^2.6.2",
+		"prettier-plugin-svelte": "^2.7.0",
+		"sass": "^1.55.0",
+		"svelte": "^3.44.0",
+		"svelte-check": "^2.7.1",
+		"svelte-preprocess": "^4.10.6",
+		"tslib": "^2.3.1",
+		"typescript": "^4.7.4",
+		"vite": "^3.1.0"
+	},
+	"type": "module",
+	"_removedDeps": {
+		"async-hydrus.js": "dpongimo/async-hydrus.js"
+	}
 }
