@@ -1,12 +1,15 @@
-const sveltePreprocess = require('svelte-preprocess');
+import adapter from "@sveltejs/adapter-static";
+import preprocess from "svelte-preprocess";
 
-module.exports = {
-  preprocess: sveltePreprocess({
-    scss: {
-      includePaths: ['src'],
-    },
-    postcss: {
-      plugins: [require('autoprefixer')],
-    },
-  }),
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	// Consult https://github.com/sveltejs/svelte-preprocess
+	// for more information about preprocessors
+	preprocess: preprocess(),
+
+	kit: {
+		adapter: adapter()
+	}
 };
+
+export default config;
