@@ -1,23 +1,25 @@
-import resolve from '@rollup/plugin-node-resolve';
-import replace from '@rollup/plugin-replace';
-import commonjs from '@rollup/plugin-commonjs';
-import svelte from 'rollup-plugin-svelte';
-import sveltePreprocess from 'svelte-preprocess';
-import babel from '@rollup/plugin-babel';
-import { terser } from 'rollup-plugin-terser';
-import config from 'sapper/config/rollup.js';
-import pkg from './package.json';
+import resolve from "@rollup/plugin-node-resolve";
+import replace from "@rollup/plugin-replace";
+import commonjs from "@rollup/plugin-commonjs";
+import svelte from "rollup-plugin-svelte";
+import sveltePreprocess from "svelte-preprocess";
+import babel from "@rollup/plugin-babel";
+import { terser } from "rollup-plugin-terser";
+import config from "sapper/config/rollup.js";
+import pkg from "./package.json";
 
 const mode = process.env.NODE_ENV;
-const dev = mode === 'development';
+const dev = mode === "development";
 const legacy = !!process.env.SAPPER_LEGACY_BUILD;
 
-const onwarn = (warning, onwarn) => (warning.code === 'CIRCULAR_DEPENDENCY' && /[/\\]@sapper[/\\]/.test(warning.message)) || onwarn(warning);
-const dedupe = importee => importee === 'svelte' || importee.startsWith('svelte/');
+const onwarn = (warning, onwarn) =>
+	(warning.code === "CIRCULAR_DEPENDENCY" && /[/\\]@sapper[/\\]/.test(warning.message)) ||
+	onwarn(warning);
+const dedupe = (importee) => importee === "svelte" || importee.startsWith("svelte/");
 
 const preprocess = sveltePreprocess({
 	scss: {
-		includePaths: ['src'],
+		includePaths: ["src"]
 	}
 });
 
@@ -27,8 +29,8 @@ export default {
 		output: config.client.output(),
 		plugins: [
 			replace({
-				'process.browser': true,
-				'process.env.NODE_ENV': JSON.stringify(mode)
+				"process.browser": true,
+				"process.env.NODE_ENV": JSON.stringify(mode)
 			}),
 			svelte({
 				preprocess,
@@ -39,36 +41,44 @@ export default {
 			resolve({
 				browser: true,
 				preferBuiltins: true,
-				dedupe: ['svelte']
+				dedupe: ["svelte"]
 			}),
 			commonjs(),
 
-			legacy && babel({
-				extensions: ['.js', '.mjs', '.html', '.svelte'],
-				babelHelpers: 'runtime',
-				exclude: ['node_modules/@babel/**'],
-				presets: [
-					['@babel/preset-env', {
-						targets: '> 0.25%, not dead'
-					}]
-				],
-				plugins: [
-					'@babel/plugin-syntax-dynamic-import',
-					['@babel/plugin-transform-runtime', {
-						useESModules: true
-					}]
-				]
-			}),
+			legacy &&
+				babel({
+					extensions: [".js", ".mjs", ".html", ".svelte"],
+					babelHelpers: "runtime",
+					exclude: ["node_modules/@babel/**"],
+					presets: [
+						[
+							"@babel/preset-env",
+							{
+								targets: "> 0.25%, not dead"
+							}
+						]
+					],
+					plugins: [
+						"@babel/plugin-syntax-dynamic-import",
+						[
+							"@babel/plugin-transform-runtime",
+							{
+								useESModules: true
+							}
+						]
+					]
+				}),
 
 			// css({ output: 'static/css/bundle.css' }),
 
-			!dev && terser({
-				module: true
-			})
+			!dev &&
+				terser({
+					module: true
+				})
 		],
 
 		preserveEntrySignatures: false,
-		onwarn,
+		onwarn
 	},
 
 	server: {
@@ -76,12 +86,12 @@ export default {
 		output: config.server.output(),
 		plugins: [
 			replace({
-				'process.browser': false,
-				'process.env.NODE_ENV': JSON.stringify(mode)
+				"process.browser": false,
+				"process.env.NODE_ENV": JSON.stringify(mode)
 			}),
 			svelte({
 				preprocess,
-				generate: 'ssr',
+				generate: "ssr",
 				dev
 			}),
 			resolve({
@@ -90,11 +100,11 @@ export default {
 			commonjs()
 		],
 		external: Object.keys(pkg.dependencies).concat(
-			require('module').builtinModules || Object.keys(process.binding('natives'))
+			require("module").builtinModules || Object.keys(process.binding("natives"))
 		),
 
-		preserveEntrySignatures: 'strict',
-		onwarn,
+		preserveEntrySignatures: "strict",
+		onwarn
 	},
 
 	serviceworker: {
@@ -103,14 +113,14 @@ export default {
 		plugins: [
 			resolve(),
 			replace({
-				'process.browser': true,
-				'process.env.NODE_ENV': JSON.stringify(mode)
+				"process.browser": true,
+				"process.env.NODE_ENV": JSON.stringify(mode)
 			}),
 			commonjs(),
 			!dev && terser()
 		],
 
 		preserveEntrySignatures: false,
-		onwarn,
+		onwarn
 	}
 };
