@@ -1,5 +1,5 @@
 <script>
-  import LightGallery from "../../components/LightGallery.svelte";
+	import LightGallery from "../../components/LightGallery.svelte";
 </script>
 
-<LightGallery></LightGallery>
+<LightGallery />
