@@ -1,5 +1,5 @@
-import { timestamp, files, shell, routes } from '@sapper/service-worker';
-import { GetClient, GetCachedAddress } from "./js/hydrus-connection.js"
+import { timestamp, files, shell, routes } from "@sapper/service-worker";
+import { GetClient, GetCachedAddress } from "./js/hydrus-connection.js";
 
 const ASSETS = `cache${timestamp}`;
 
@@ -8,20 +8,20 @@ const ASSETS = `cache${timestamp}`;
 const to_cache = shell.concat(files);
 const cached = new Set(to_cache);
 
-self.addEventListener('install', event => {
+self.addEventListener("install", (event) => {
 	event.waitUntil(
 		caches
 			.open(ASSETS)
-			.then(cache => cache.addAll(to_cache))
+			.then((cache) => cache.addAll(to_cache))
 			.then(() => {
 				self.skipWaiting();
 			})
 	);
 });
 
-self.addEventListener('activate', event => {
+self.addEventListener("activate", (event) => {
 	event.waitUntil(
-		caches.keys().then(async keys => {
+		caches.keys().then(async (keys) => {
 			// delete old caches
 			for (const key of keys) {
 				if (key !== ASSETS) await caches.delete(key);
@@ -32,14 +32,13 @@ self.addEventListener('activate', event => {
 	);
 });
 
-self.addEventListener('fetch', event => {
-	if (event.request.method !== 'GET' || event.request.headers.has('range')) return;
+self.addEventListener("fetch", (event) => {
+	if (event.request.method !== "GET" || event.request.headers.has("range")) return;
 
 	const url = new URL(event.request.url);
-	
 
 	// don't try to handle e.g. data: URIs
-	if (!url.protocol.startsWith('http')) return;
+	if (!url.protocol.startsWith("http")) return;
 
 	// ignore dev server requests
 	if (url.hostname === self.location.hostname && url.port !== self.location.port) return;
@@ -53,39 +52,35 @@ self.addEventListener('fetch', event => {
 	// for pages, you might want to serve a shell `service-worker-index.html` file,
 	// which Sapper has generated for you. It's not right for every
 	// app, but if it's right for yours then uncomment this section
-	if (url.origin === self.origin && routes.find(route => route.pattern.test(url.pathname))) {
-		event.respondWith(caches.match('/service-worker-index.html'));
+	if (url.origin === self.origin && routes.find((route) => route.pattern.test(url.pathname))) {
+		event.respondWith(caches.match("/service-worker-index.html"));
 		return;
 	}
 
-	if (event.request.cache === 'only-if-cached') return;
+	if (event.request.cache === "only-if-cached") return;
 
 	// for everything else, try the network first, falling back to
 	// cache if the user is offline. (If the pages never change, you
 	// might prefer a cache-first approach to a network-first one.)
 	event.respondWith(
-		caches
-			.open(`offline${timestamp}`)
-			.then(async cache => {
-				// Reply with cached files immediately as file request urls are immutable
-				if (url.pathname.includes("get_files/file")) {
-					const response = await cache.match(event.request);
-					if (response) return response;
-				}
-				try {
-					const response = await fetch(event.request);
-					cache.put(event.request, response.clone());
-					return response;
-				} catch (err) {
-					const response = await cache.match(event.request);
-					if (response) return response;
+		caches.open(`offline${timestamp}`).then(async (cache) => {
+			// Reply with cached files immediately as file request urls are immutable
+			if (url.pathname.includes("get_files/file")) {
+				const response = await cache.match(event.request);
+				if (response) return response;
+			}
+			try {
+				const response = await fetch(event.request);
+				cache.put(event.request, response.clone());
+				return response;
+			} catch (err) {
+				const response = await cache.match(event.request);
+				if (response) return response;
 
-					throw err;
-				}
-			})
+				throw err;
+			}
+		})
 	);
 });
 
-addEventListener('message', event => {
-
-});
+addEventListener("message", (event) => {});
